Fix distribution detection in bootstrap
The bootstrap-ansible script is responsible for installing many base
packages. It attempts to detect the host distribution by checking for
various operating system identification files (such as /etc/os-release).

As per the freedesktop.org spec, all complying distributions ship the
/etc/os-release file, just with different content. The script
incorrectly assumes that if this file is present it represents an apt
based distribution, such as Ubuntu, and proceeds to install base
packages. As all distributions ship this file, the bootstrap process
fails on non-apt systems, such as Fedora (the subsequent checks for
distribution-specific files, such as /etc/fedora-release, are skipped).

This patch adds a function to the scripts-library which will read the
/etc/os-release file and export various DISTRO_ prefixed variables which
identify the distribution and version. These can be used wherever the
scripts-library is sourced.

The package installation check inside bootstrap-ansible.sh is updated
to make use of these new variables by checking the value of DISTRO_ID.

Related to commits:
d9b9a427aa03 "Add reliable OS detection for ansible bootstrap"
68d68c27dd80 "Remove os-detection script"

Implements: blueprint multi-platform-host

Change-Id: I07c8a7a7f8d4cf56cf125fcef18e6cf4f473e39e