Fix override of ANSIBLE_PACKAGE variable
The initialisation of the ANSIBLE_PACKAGE variable in bootstrap-ansible
should be overridden by any existing environment variable of the same
name, however it is incorrectly looking for "ANSIBLE_RELEASE." This means
that any attempt to override the Ansible version will be ignored.

This patch fixes it so that the script looks for a variable of the same
name, "ANSIBLE_PACKAGE," and also updates the documentation.

Related to commit:
88ae508e0f95 "Install Ansible from pypi instead of from a git clone."

Change-Id: I5cf40ea855f05073fbe568cd9839dbd5134e19bf