Robust base dir support for bootstrapping
Currently the library expects the user to be running bootstrap scripts
from the root of the openstack-ansible directory. If the user is not,
then the bootstrap exists with failure and a warning.

Directories with special characters (such as spaces) are also not
supported, so running the script from outside the base dir will cause
script errors.

This patch implements support for bootstrapping outside the base dir,
including support or directory names with special characters. This is
extended to test and upgrade scripts.

Change-Id: I5cfa5a0bdbd762c50fe5a41cb88b3c0677f62482